@@ -12,15 +12,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.9",
-    "@angular/common": "6.0.9",
-    "@angular/compiler": "6.0.9",
-    "@angular/core": "6.0.9",
-    "@angular/forms": "6.0.9",
-    "@angular/http": "6.0.9",
-    "@angular/platform-browser": "6.0.9",
-    "@angular/platform-browser-dynamic": "6.0.9",
-    "@angular/router": "6.0.9",
+    "@angular/animations": "8.2.4",
+    "@angular/common": "8.2.4",
+    "@angular/compiler": "8.2.4",
+    "@angular/core": "8.2.4",
+    "@angular/forms": "8.2.4",
+    "@angular/platform-browser": "8.2.4",
+    "@angular/platform-browser-dynamic": "8.2.4",
+    "@angular/router": "8.2.4",
     "aos": "^2.2.0",
     "bootstrap": "4.1.2",
     "contentful": "^6.1.1",
@@ -28,24 +27,24 @@
     "jquery": "3.2.1",
     "marked": "^0.3.19",
     "moment": "^2.22.2",
-    "ngx-md": "^3.1.1",
+    "ngx-md": "^8.0.0",
     "popper.js": "^1.14.1",
     "request": "^2.87.0",
-    "rxjs": "6.2.2",
+    "rxjs": "6.5.2",
     "rxjs-compat": "6.2.2",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "^6.2.6",
-    "@angular/compiler-cli": "6.0.9",
-    "@angular/language-service": "6.0.9",
+    "@angular-devkit/build-angular": "^0.803.2",
+    "@angular/cli": "^8.3.2",
+    "@angular/compiler-cli": "8.2.4",
+    "@angular/language-service": "8.2.4",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/jquery": "^3.2.7",
     "@types/marked": "^0.3.0",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^5.1.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.0.0",
@@ -56,6 +55,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "2.7.2"
+    "typescript": "3.5.3"
   }
 }
